All supporters pages
Supporter pages use the worldmap plugin.
Generatemap in general js is used to generate the map. colorSuppMap is
used to color the top 5 countries and generate the legend next to the
map. The codes for the top 5 countries are in the global var countries,
and all country codes used by the map plugin can be found in
worldmap-mill.js.